requirements.txt
python-can==3.3.4
pyserial==3.5
cantools==35.5.0
